backend/pyproject.toml
[project]
name = "bobby"
version = "0.1.0"
description = "Application Bobby - Plateforme de cooptation ESN"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    "redis>=5.2.0",
    "aiosmtplib>=3.0.2",
    "resend>=2.5.0",
    "structlog>=24.4.0",
    "python-multipart>=0.0.18",
    "email-validator>=2.2.0",
    "google-generativeai>=0.8.3",
    "anthropic>=0.40.0",
    "docxtpl>=0.19.0",
    "python-docx>=1.1.2",
    "openpyxl>=3.1.5",
    "PyPDF2>=3.0.1",
    "aioboto3>=13.2.0",
    "boto3>=1.35.0",
    # Security (rate limiting)
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-watch>=4.2.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "factory-boy>=3.3.1",
    "faker>=33.0.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise-without-from-inside-except
    "E712",  # true-false-comparison
    "E402",  # module-import-not-at-top-of-file
    "UP042", # replace-str-enum
    "F841",  # unused-variable
    "UP046", # non-pep695-generic-class
    "B905",  # zip-without-explicit-strict
    "B007",  # unused-loop-control-variable
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]
